project cleanup
removed unnecessary code, comments, and files

removed tests and factories as these are beyond the scope of this project

removed unused components, including unused layout components

created the page layout component and applied it to all pages. this reduced the amount of repeated code

restructured the Pages directory to be more intuitive. restructured other directories as authentication is the main feature of this application